pyproject.toml
[project]
name = "aug-sdk-demo"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Tim Misiak", email = "tim.misiak+github@gmail.com" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "aug-sdk @ git+https://github.com/AugmendTech/aug-sdk",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
aug-sdk-demo = "aug_sdk_demo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/aug_sdk_demo"]
